refactor(Utilities): Put validate and coerce into separate files
